+++ plugins/flash/dialogs/flash.js
@@ -87,7 +87,11 @@
 	}, {
 		type: ATTRTYPE_PARAM, name: names[ i ]
 	} ];
-	names = [ 'allowFullScreen', 'play', 'loop', 'menu' ];
+
+	// These attributes are "true" by default and not present in editor data (when "true").
+	// Note that, though default value of "allowFullScreen" is "true", it is not listed here.
+	// "allowFullScreen" is present in editor data regardless of the value (#7634).
+	names = [ 'play', 'loop', 'menu' ];
 	for ( i = 0; i < names.length; i++ )
 		attributesMap[ names[ i ] ][ 0 ][ 'default' ] = attributesMap[ names[ i ] ][ 1 ][ 'default' ] = true;